generate-stackbrew-library: Enable AArch64
By default the Docker Images project only builds and pushes (to
Docker Hub) images for AMD64.  In order to support additional
architectures the 'Architectures:' label is used to denote which
ones.

AArch64 has been tested working on all currently supported
versions of YOURLS.  This means we can make the 'Architecture'
global in the library file which simplifies things quite a bit.

Signed-off-by: Lee Jones <lee.jones@linaro.org>